bot/requirements.txt
SQLAlchemy==1.3.22
aiohttp==3.7.4
PythonTwitchBotFramework==2.1.10
pytz==2020.4
aiofile==3.3.3
Adafruit_IO==2.4.0
alembic==1.4.3
mysql-connector-python==8.0.22
gmqtt==0.6.9
dropbox
